requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
cloudinary==1.32.0
distro==1.9.0
dj-database-url==0.5.0
dj-rest-auth==5.0.2
dj3-cloudinary-storage==0.0.6
Django==5.0.3
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-filter==24.1
django-jazzmin==2.6.1
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==20.1.0
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.6
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==2.0.0
sqlparse==0.4.4
tqdm==4.66.2
urllib3==1.26.18
whitenoise==6.6.0
yarl==1.9.4
